Made Storage Class Folder
Also, improved comments explaining storage class functions
